fix: add path enums, prettier ignore + tabWidth
