+++ src/App.vue
@@ -1,9 +1,9 @@
 <template>
-    <v-app>
-        <v-main>
-            <router-view />
-        </v-main>
-    </v-app>
+  <v-app>
+    <v-main>
+      <router-view />
+    </v-main>
+  </v-app>
 </template>
 
 <script lang="ts" setup></script>
