+++ src/main.ts
@@ -1,6 +1,5 @@
 import { registerPlugins } from '@/plugins';
 import App from './App.vue';
-import { createApp } from 'vue';
 
 const app = createApp(App);
 registerPlugins(app);
