+++ src/modules/auth/enums/auth-path-names.enum.ts
@@ -1,5 +1,5 @@
-export enum AuthRoutesNames {
-  AUTH = 'auth',
+export enum AuthPathNames {
+  ROOT = '',
   LOGIN = 'login',
   FORGOT_PASSWORD = 'forgot-password',
 }
