+++ src/modules/auth/views/login-view.vue
@@ -1,3 +1,3 @@
 <template>
-    <div>Login?</div>
+  <div>Login?</div>
 </template>
