+++ src/modules/dashboard/dashboard-view.vue
@@ -1,3 +1,3 @@
 <template>
-    <div>/dashboard</div>
+  <div>/dashboard</div>
 </template>
