+++ src/plugins/index.ts
@@ -4,5 +4,5 @@ import router from './router';
 import type { App } from 'vue';
 
 export function registerPlugins(app: App) {
-    app.use(vuetify).use(router).use(pinia);
+  app.use(vuetify).use(router).use(pinia);
 }
