+++ src/plugins/router.ts
@@ -1,47 +1,47 @@
 import { createRouter, createWebHistory } from 'vue-router/auto';
 import { authRoutes } from '@/modules/auth/routes';
-import { dashboardRoutes } from '@/modules/dashboard/dashboard.routes';
+import { dashboardRoutes } from '@/modules/dashboard/routes';
 
 const routes = [
-    {
-        path: '/',
-        children: [...dashboardRoutes],
-    },
-    {
-        path: '/auth',
-        children: [...authRoutes],
-    },
+  {
+    path: '/',
+    children: [...dashboardRoutes],
+  },
+  {
+    path: '/auth',
+    children: [...authRoutes],
+  },
 ];
 
 const router = createRouter({
-    history: createWebHistory(import.meta.env.BASE_URL),
-    routes,
+  history: createWebHistory(import.meta.env.BASE_URL),
+  routes,
 });
 
 // Workaround for https://github.com/vitejs/vite/issues/11804
 router.onError((err, to) => {
-    if (
-        err?.message?.includes?.(
-            'Failed to fetch dynamically imported module',
-        )
-    ) {
-        if (!localStorage.getItem('vuetify:dynamic-reload')) {
-            console.log('Reloading page to fix dynamic import error');
-            localStorage.setItem('vuetify:dynamic-reload', 'true');
-            location.assign(to.fullPath);
-        } else {
-            console.error(
-                'Dynamic import error, reloading page did not fix it',
-                err,
-            );
-        }
+  if (
+    err?.message?.includes?.(
+      'Failed to fetch dynamically imported module',
+    )
+  ) {
+    if (!localStorage.getItem('vuetify:dynamic-reload')) {
+      console.log('Reloading page to fix dynamic import error');
+      localStorage.setItem('vuetify:dynamic-reload', 'true');
+      location.assign(to.fullPath);
     } else {
-        console.error(err);
+      console.error(
+        'Dynamic import error, reloading page did not fix it',
+        err,
+      );
     }
+  } else {
+    console.error(err);
+  }
 });
 
 router.isReady().then(() => {
-    localStorage.removeItem('vuetify:dynamic-reload');
+  localStorage.removeItem('vuetify:dynamic-reload');
 });
 
 export default router;
