+++ src/plugins/vuetify.ts
@@ -1,19 +1,9 @@
-/**
- * plugins/vuetify.ts
- *
- * Framework documentation: https://vuetifyjs.com`
- */
-
-// Styles
 import '@mdi/font/css/materialdesignicons.css';
 import 'vuetify/styles';
-
-// Composables
 import { createVuetify } from 'vuetify';
 
-// https://vuetifyjs.com/en/introduction/why-vuetify/#feature-guides
 export default createVuetify({
-    theme: {
-        defaultTheme: 'dark',
-    },
+  theme: {
+    defaultTheme: 'dark',
+  },
 });
